% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_watch.R
\name{.anchors}
\alias{.anchors}
\title{Links in an HTML fragment}
\usage{
.anchors(html)
}
\value{
data.table(href, text) with entities decoded and tags stripped.
}
\description{
Links in an HTML fragment
}
\keyword{internal}
