% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{.catalog_table}
\alias{.catalog_table}
\title{The variable key}
\usage{
.catalog_table()
}
\value{
data.table with dataset, variable, description, counting.
}
\description{
The variable key
}
\keyword{internal}
