% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_augmented}
\alias{.check_augmented}
\title{Checks run at the end of augment_analytic()}
\usage{
.check_augmented(at, ls_years, xwalk_years)
}
\description{
Checks run at the end of augment_analytic()
}
\keyword{internal}
