% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_dec_new_plans}
\alias{.check_dec_new_plans}
\title{Check that plans new in landscape year Y have December Y enrollment}
\usage{
.check_dec_new_plans(dfl, dfe_dec_year, dec_year)
}
\description{
Historically 88-94\% do; a December file that is really the previous
year's scores near 0\%.
}
\keyword{internal}
