% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_new_plans}
\alias{.check_new_plans}
\title{Check that the crosswalk's new MA plans appear in January enrollment}
\usage{
.check_new_plans(dfnew, dfe_jan_year, xwalk_year)
}
\description{
Historically 77-92\% of new MA plans have January enrollment.
}
\keyword{internal}
