% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.check_new_plans_landscape}
\alias{.check_new_plans_landscape}
\title{Check that the crosswalk's new MA plans appear in landscape year N}
\usage{
.check_new_plans_landscape(dfx, dfl, xwalk_year)
}
\description{
Historically 98-100\% do. A low share means the landscape registered for
year N is really another year's file.
}
\keyword{internal}
