% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_watch.R
\name{.downloads_block}
\alias{.downloads_block}
\title{The "Downloads" list of a CMS page (NA if the page has none)}
\usage{
.downloads_block(html)
}
\description{
The "Downloads" list of a CMS page (NA if the page has none)
}
\keyword{internal}
