% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.R
\name{.get_landscape_2024}
\alias{.get_landscape_2024}
\title{Read the CY2024 landscape (separate MA + SNP files, plus sanctioned plans)}
\usage{
.get_landscape_2024()
}
\description{
Read the CY2024 landscape (separate MA + SNP files, plus sanctioned plans)
}
\keyword{internal}
