% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.R
\name{.landscape_2024_paths}
\alias{.landscape_2024_paths}
\title{Paths of the four CY2024 landscape files}
\usage{
.landscape_2024_paths()
}
\description{
Paths of the four CY2024 landscape files
}
\keyword{internal}
