% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.R
\name{.landscape_file}
\alias{.landscape_file}
\title{Path of the registered CY2025+ combined landscape file}
\usage{
.landscape_file(cy)
}
\description{
Path of the registered CY2025+ combined landscape file
}
\keyword{internal}
