% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{.parse_cpsc_name}
\alias{.parse_cpsc_name}
\title{Parse year and month from a CPSC_Enrollment_Info_YYYY_MM.csv file name}
\usage{
.parse_cpsc_name(path)
}
\description{
Parse year and month from a CPSC_Enrollment_Info_YYYY_MM.csv file name
}
\keyword{internal}
