% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_details.R
\name{.parse_star}
\alias{.parse_star}
\title{Parse star ratings ("4 Stars", "3.5", "4.0") to numbers; labels are NA}
\usage{
.parse_star(x)
}
\description{
Parse star ratings ("4 Stars", "3.5", "4.0") to numbers; labels are NA
}
\keyword{internal}
