% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.plan_detail_years}
\alias{.plan_detail_years}
\title{Contract years with plan details for a set of crosswalk years}
\usage{
.plan_detail_years(xwalk_years = MAEXITS_XWALK_YEARS)
}
\description{
Contract years with plan details for a set of crosswalk years
}
\keyword{internal}
