% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_watch.R
\name{.plist_xml}
\alias{.plist_xml}
\title{launchd property list for the daily check}
\usage{
.plist_xml(program, hour, minute, log, workdir)
}
\description{
launchd property list for the daily check
}
\keyword{internal}
