% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_details.R
\name{.star_status}
\alias{.star_status}
\title{Star rating status: rated, not enough data, plan too new, not applicable}
\usage{
.star_status(x)
}
\description{
Star rating status: rated, not enough data, plan too new, not applicable
}
\keyword{internal}
