% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_augment.R
\name{.state_name_from_abb}
\alias{.state_name_from_abb}
\title{Full state or territory name from a postal abbreviation}
\usage{
.state_name_from_abb(abb)
}
\description{
Full state or territory name from a postal abbreviation
}
\keyword{internal}
