% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.statuses}
\alias{.statuses}
\title{Crosswalk statuses belonging to one or more classes}
\usage{
.statuses(cls)
}
\arguments{
\item{cls}{Character vector of classes from
\code{MAEXITS_XWALK_STATUS_CLASS} ("continuing",
"service_area_reduction", "service_area_expansion", "terminated", "new").}
}
\value{
Character vector of status strings.
}
\description{
Crosswalk statuses belonging to one or more classes
}
\keyword{internal}
